Visibility more or less working, some rounding bugs. Changed point to be floating pt value which fixed a lot of problems

+++ src/ShortPath/GJK.java
@@ -18,7 +18,7 @@ public class GJK {
         Point closestPoint=polygon.getHull().get(0);
         double maxDotProduct=dotProduct(polygon.getHull().get(0),direction);
         for(Point P:polygon.getHull()){
-            int dotProduct=P.getX()*direction.getX()+P.getY()*direction.getY();
+            float dotProduct=P.getX()*direction.getX()+P.getY()*direction.getY();
             if (dotProduct>maxDotProduct){
                 maxDotProduct=dotProduct;
                 closestPoint=P;
